BAH-4858 | Refactor. Move getObservationsBundleByEncounterUuid to observationService instead of encounterService
