+++ packages/bahmni-services/src/observationService/index.ts
@@ -2,4 +2,5 @@ export {
   getPatientObservationsBundle,
   getPatientObservationsWithEncounterBundle,
   getPatientObservations,
+  getObservationsBundleByEncounterUuid,
 } from './observationService';
